i have added a table in addSales tab and some validation methods to validate emails and phone number

+++ WholesaleInventorySystem/src/wholesaleinventorysystem/Products.java
@@ -10,6 +10,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.Optional;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javafx.collections.FXCollections;
@@ -17,7 +18,9 @@ import javafx.collections.ObservableList;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
 import javafx.scene.control.Tab;
@@ -150,6 +153,14 @@ public class Products  {
         
         deleteButton.setStyle("-fx-text-fill:white;");
         deleteButton.setOnAction(e->{
+           Alert confAlert = new Alert(Alert.AlertType.CONFIRMATION);
+           confAlert.setTitle("Confirmation dialog");
+           confAlert.setHeaderText(null);
+           confAlert.setContentText("Are sure you want to deleted the item?");
+            
+           Optional<ButtonType> action=confAlert.showAndWait();
+           if(action.get()==ButtonType.OK){
+                
             
             try {
                 String query="DELETE FROM product where ProductId=?";
@@ -167,6 +178,7 @@ public class Products  {
             } 
             catch (SQLException ex) {
                 Logger.getLogger(Products.class.getName()).log(Level.SEVERE, null, ex);
+            }
             }
         });
